introducing c++ casts
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk/qwt@1071 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ examples/plotmatrix/plotmatrix.cpp
@@ -84,7 +84,7 @@ int PlotMatrix::numColumns() const
 QwtPlot* PlotMatrix::plot(int row, int column)
 {
     const int index = row * numColumns() + column;
-    if ( index < (int)d_data->plotWidgets.size() )
+    if ( index < d_data->plotWidgets.size() )
         return d_data->plotWidgets[index];
 
     return NULL;
@@ -93,7 +93,7 @@ QwtPlot* PlotMatrix::plot(int row, int column)
 const QwtPlot* PlotMatrix::plot(int row, int column) const
 {
     const int index = row * numColumns() + column;
-    if ( index < (int)d_data->plotWidgets.size() )
+    if ( index < d_data->plotWidgets.size() )
         return d_data->plotWidgets[index];
 
     return NULL;
